+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigtable/BigtableIO.java
@@ -17,6 +17,7 @@
  */
 package org.apache.beam.sdk.io.gcp.bigtable;
 
+import static org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkArgument;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkNotNull;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkState;
@@ -30,7 +31,6 @@ import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.protobuf.ByteString;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
@@ -207,12 +207,7 @@ public class BigtableIO {
 
     abstract BigtableConfig getBigtableConfig();
 
-    @Nullable
-    abstract RowFilter getRowFilter();
-
-    /** Returns the range of keys that will be read from the table. */
-    @Nullable
-    public abstract List<ByteKeyRange> getKeyRanges();
+    abstract BigtableReadOptions getBigtableReadOptions();
 
     /** Returns the table being read from. */
     @Nullable
@@ -237,13 +232,15 @@ public class BigtableIO {
     static Read create() {
       BigtableConfig config =
           BigtableConfig.builder()
-              .setTableId(ValueProvider.StaticValueProvider.of(""))
+              .setTableId(StaticValueProvider.of(""))
               .setValidate(true)
               .build();
 
       return new AutoValue_BigtableIO_Read.Builder()
           .setBigtableConfig(config)
-          .setKeyRanges(Arrays.asList(ByteKeyRange.ALL_KEYS))
+          .setBigtableReadOptions(BigtableReadOptions.builder()
+              .setKeyRanges(StaticValueProvider
+                  .of(Collections.singletonList(ByteKeyRange.ALL_KEYS))).build())
           .build();
     }
 
@@ -252,9 +249,7 @@ public class BigtableIO {
 
       abstract Builder setBigtableConfig(BigtableConfig bigtableConfig);
 
-      abstract Builder setRowFilter(RowFilter filter);
-
-      abstract Builder setKeyRanges(List<ByteKeyRange> keyRange);
+      abstract Builder setBigtableReadOptions(BigtableReadOptions bigtableReadOptions);
 
       abstract Read build();
     }
@@ -279,7 +274,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Read withProjectId(String projectId) {
-      return withProjectId(ValueProvider.StaticValueProvider.of(projectId));
+      return withProjectId(StaticValueProvider.of(projectId));
     }
 
     /**
@@ -302,7 +297,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Read withInstanceId(String instanceId) {
-      return withInstanceId(ValueProvider.StaticValueProvider.of(instanceId));
+      return withInstanceId(StaticValueProvider.of(instanceId));
     }
 
     /**
@@ -321,7 +316,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Read withTableId(String tableId) {
-      return withTableId(ValueProvider.StaticValueProvider.of(tableId));
+      return withTableId(StaticValueProvider.of(tableId));
     }
 
     /**
@@ -389,9 +384,22 @@ public class BigtableIO {
      *
      * <p>Does not modify this object.
      */
-    public Read withRowFilter(RowFilter filter) {
+    public Read withRowFilter(ValueProvider<RowFilter> filter) {
       checkArgument(filter != null, "filter can not be null");
-      return toBuilder().setRowFilter(filter).build();
+      BigtableReadOptions bigtableReadOptions = getBigtableReadOptions();
+      return toBuilder()
+          .setBigtableReadOptions(bigtableReadOptions.toBuilder().setRowFilter(filter).build())
+          .build();
+    }
+
+    /**
+     * Returns a new {@link BigtableIO.Read} that will filter the rows read from Cloud Bigtable
+     * using the given row filter.
+     *
+     * <p>Does not modify this object.
+     */
+    public Read withRowFilter(RowFilter filter) {
+      return withRowFilter(StaticValueProvider.of(filter));
     }
 
     /**
@@ -400,8 +408,21 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Read withKeyRange(ByteKeyRange keyRange) {
-      checkArgument(keyRange != null, "keyRange can not be null");
-      return toBuilder().setKeyRanges(Arrays.asList(keyRange)).build();
+      return withKeyRanges(Collections.singletonList(keyRange));
+    }
+
+    /**
+     * Returns a new {@link BigtableIO.Read} that will read only rows in the specified ranges.
+     * Ranges must not overlap.
+     *
+     * <p>Does not modify this object.
+     */
+    public Read withKeyRanges(ValueProvider<List<ByteKeyRange>> keyRanges) {
+      checkArgument(keyRanges != null, "keyRanges can not be null");
+      BigtableReadOptions bigtableReadOptions = getBigtableReadOptions();
+      return toBuilder()
+          .setBigtableReadOptions(bigtableReadOptions.toBuilder().setKeyRanges(keyRanges).build())
+          .build();
     }
 
     /**
@@ -411,12 +432,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Read withKeyRanges(List<ByteKeyRange> keyRanges) {
-      checkArgument(keyRanges != null, "keyRanges can not be null");
-      checkArgument(!keyRanges.isEmpty(), "keyRanges can not be empty");
-      for (ByteKeyRange range : keyRanges) {
-        checkArgument(range != null, "keyRanges cannot hold null range");
-      }
-      return toBuilder().setKeyRanges(keyRanges).build();
+      return withKeyRanges(StaticValueProvider.of(keyRanges));
     }
 
     /** Disables validation that the table being read from exists. */
@@ -442,9 +458,10 @@ public class BigtableIO {
     @Override
     public PCollection<Row> expand(PBegin input) {
       getBigtableConfig().validate();
+      getBigtableReadOptions().validate();
 
       BigtableSource source =
-          new BigtableSource(getBigtableConfig(), getRowFilter(), getKeyRanges(), null);
+          new BigtableSource(getBigtableConfig(), getBigtableReadOptions(), null);
       return input.getPipeline().apply(org.apache.beam.sdk.io.Read.from(source));
     }
 
@@ -457,28 +474,14 @@ public class BigtableIO {
     public void populateDisplayData(DisplayData.Builder builder) {
       super.populateDisplayData(builder);
       getBigtableConfig().populateDisplayData(builder);
-
-      List<ByteKeyRange> keyRanges = getKeyRanges();
-      for (int i = 0; i < keyRanges.size() && i < 5; i++) {
-        builder.addIfNotDefault(
-            DisplayData.item("keyRange " + i, keyRanges.get(i).toString()),
-            ByteKeyRange.ALL_KEYS.toString());
-      }
-
-      if (getRowFilter() != null) {
-        builder.add(
-            DisplayData.item("rowFilter", getRowFilter().toString()).withLabel("Table Row Filter"));
-      }
+      getBigtableReadOptions().populateDisplayData(builder);
     }
 
     @Override
     public String toString() {
       ToStringHelper helper =
           MoreObjects.toStringHelper(Read.class).add("config", getBigtableConfig());
-      for (int i = 0; i < getKeyRanges().size(); i++) {
-        helper.add("keyRange " + i, getKeyRanges().get(i));
-      }
-      return helper.add("filter", getRowFilter()).toString();
+      return helper.add("readOptions", getBigtableReadOptions()).toString();
     }
   }
 
@@ -524,7 +527,7 @@ public class BigtableIO {
     static Write create() {
       BigtableConfig config =
           BigtableConfig.builder()
-              .setTableId(ValueProvider.StaticValueProvider.of(""))
+              .setTableId(StaticValueProvider.of(""))
               .setValidate(true)
               .setBigtableOptionsConfigurator(enableBulkApiConfigurator(null))
               .build();
@@ -560,7 +563,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Write withProjectId(String projectId) {
-      return withProjectId(ValueProvider.StaticValueProvider.of(projectId));
+      return withProjectId(StaticValueProvider.of(projectId));
     }
 
     /**
@@ -583,7 +586,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Write withInstanceId(String instanceId) {
-      return withInstanceId(ValueProvider.StaticValueProvider.of(instanceId));
+      return withInstanceId(StaticValueProvider.of(instanceId));
     }
 
     /**
@@ -602,7 +605,7 @@ public class BigtableIO {
      * <p>Does not modify this object.
      */
     public Write withTableId(String tableId) {
-      return withTableId(ValueProvider.StaticValueProvider.of(tableId));
+      return withTableId(StaticValueProvider.of(tableId));
     }
 
     /**
@@ -870,12 +873,10 @@ public class BigtableIO {
   static class BigtableSource extends BoundedSource<Row> {
     public BigtableSource(
         BigtableConfig config,
-        @Nullable RowFilter filter,
-        List<ByteKeyRange> ranges,
+        BigtableReadOptions readOptions,
         @Nullable Long estimatedSizeBytes) {
       this.config = config;
-      this.filter = filter;
-      this.ranges = ranges;
+      this.readOptions = readOptions;
       this.estimatedSizeBytes = estimatedSizeBytes;
     }
 
@@ -883,28 +884,26 @@ public class BigtableIO {
     public String toString() {
       return MoreObjects.toStringHelper(BigtableSource.class)
           .add("config", config)
-          .add("filter", filter)
-          .add("ranges", ranges)
+          .add("readOptions", readOptions)
           .add("estimatedSizeBytes", estimatedSizeBytes)
           .toString();
     }
 
     ////// Private state and internal implementation details //////
     private final BigtableConfig config;
-    @Nullable private final RowFilter filter;
-    private final List<ByteKeyRange> ranges;
+    private final BigtableReadOptions readOptions;
     @Nullable private Long estimatedSizeBytes;
     @Nullable private transient List<SampleRowKeysResponse> sampleRowKeys;
 
     /** Creates a new {@link BigtableSource} with just one {@link ByteKeyRange}. */
     protected BigtableSource withSingleRange(ByteKeyRange range) {
       checkArgument(range != null, "range can not be null");
-      return new BigtableSource(config, filter, Arrays.asList(range), estimatedSizeBytes);
+      return new BigtableSource(config, readOptions.withKeyRange(range), estimatedSizeBytes);
     }
 
     protected BigtableSource withEstimatedSizeBytes(Long estimatedSizeBytes) {
       checkArgument(estimatedSizeBytes != null, "estimatedSizeBytes can not be null");
-      return new BigtableSource(config, filter, ranges, estimatedSizeBytes);
+      return new BigtableSource(config, readOptions, estimatedSizeBytes);
     }
 
     /**
@@ -957,7 +956,8 @@ public class BigtableIO {
       for (BigtableSource source : splits) {
         if (counter == numberToCombine
             || !checkRangeAdjacency(previousSourceRanges, source.getRanges())) {
-          reducedSplits.add(new BigtableSource(config, filter, previousSourceRanges, size));
+          reducedSplits.add(
+              new BigtableSource(config, readOptions.withKeyRanges(previousSourceRanges), size));
           counter = 0;
           size = 0;
           previousSourceRanges = new ArrayList<>();
@@ -968,7 +968,7 @@ public class BigtableIO {
         counter++;
       }
       if (size > 0) {
-        reducedSplits.add(new BigtableSource(config, filter, previousSourceRanges, size));
+        reducedSplits.add(new BigtableSource(config, readOptions.withKeyRanges(previousSourceRanges), size));
       }
       return reducedSplits;
     }
@@ -1042,7 +1042,7 @@ public class BigtableIO {
           sampleRowKeys.get(0));
 
       ImmutableList.Builder<BigtableSource> splits = ImmutableList.builder();
-      for (ByteKeyRange range : ranges) {
+      for (ByteKeyRange range : getRanges()) {
         splits.addAll(splitRangeBasedOnSamples(desiredBundleSizeBytes, sampleRowKeys, range));
       }
       return splits.build();
@@ -1149,7 +1149,7 @@ public class BigtableIO {
           lastOffset = currentOffset;
           continue;
         } else {
-          for (ByteKeyRange range : ranges) {
+          for (ByteKeyRange range : getRanges()) {
             if (range.overlaps(ByteKeyRange.of(currentStartKey, currentEndKey))) {
               estimatedSizeBytes += currentOffset - lastOffset;
               // We don't want to double our estimated size if two ranges overlap this sample
@@ -1188,8 +1188,8 @@ public class BigtableIO {
 
       builder.add(DisplayData.item("tableId", config.getTableId()).withLabel("Table ID"));
 
-      if (filter != null) {
-        builder.add(DisplayData.item("rowFilter", filter.toString()).withLabel("Table Row Filter"));
+      if (getRowFilter() != null) {
+        builder.add(DisplayData.item("rowFilter", getRowFilter().toString()).withLabel("Table Row Filter"));
       }
     }
 
@@ -1234,11 +1234,13 @@ public class BigtableIO {
     }
 
     public List<ByteKeyRange> getRanges() {
-      return ranges;
+      return readOptions.getKeyRanges().get();
     }
 
+    @Nullable
     public RowFilter getRowFilter() {
-      return filter;
+      ValueProvider<RowFilter> rowFilter = readOptions.getRowFilter();
+      return rowFilter != null && rowFilter.isAccessible() ? rowFilter.get() : null;
     }
 
     public ValueProvider<String> getTableId() {