+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/bigtable/BigtableIOTest.java
@@ -17,6 +17,7 @@
  */
 package org.apache.beam.sdk.io.gcp.bigtable;
 
+import static org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
 import static org.apache.beam.sdk.testing.SourceTestUtils.assertSourcesEqualReferenceSource;
 import static org.apache.beam.sdk.testing.SourceTestUtils.assertSplitAtFractionExhaustive;
 import static org.apache.beam.sdk.testing.SourceTestUtils.assertSplitAtFractionFails;
@@ -59,6 +60,7 @@ import java.io.Serializable;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -182,6 +184,9 @@ public class BigtableIOTest {
 
   private static final SerializableFunction<BigtableOptions.Builder, BigtableOptions.Builder>
       PORT_CONFIGURATOR = input -> input.setPort(1234);
+
+  private static final ValueProvider<List<ByteKeyRange>> ALL_KEY_RANGE = StaticValueProvider
+      .of(Collections.singletonList(ByteKeyRange.ALL_KEYS));
 
   @Before
   public void setup() throws Exception {
@@ -513,7 +518,7 @@ public class BigtableIOTest {
   }
 
   private void runReadTest(BigtableIO.Read read, List<Row> expected) {
-    PCollection<Row> rows = p.apply(read.getTableId() + "_" + read.getKeyRanges(), read);
+    PCollection<Row> rows = p.apply(read.getTableId() + "_" + read.getBigtableReadOptions().getKeyRanges(), read);
     PAssert.that(rows).containsInAnyOrder(expected);
     p.run();
   }
@@ -561,6 +566,26 @@ public class BigtableIOTest {
 
     // Suffix should contain the middle.
     assertThat(suffixRows, hasItems(middleRows.toArray(new Row[] {})));
+  }
+
+  /**
+   * Tests reading key ranges specified through a ValueProvider.
+   */
+  @Test
+  public void testReadingWithRuntimeParameterizedKeyRange() throws Exception {
+    final String table = "TEST-KEY-RANGE-TABLE";
+    final int numRows = 1001;
+    List<Row> testRows = makeTableData(table, numRows);
+    ByteKey startKey = ByteKey.copyFrom("key000000100".getBytes(StandardCharsets.UTF_8));
+    ByteKey endKey = ByteKey.copyFrom("key000000300".getBytes(StandardCharsets.UTF_8));
+
+    service.setupSampleRowKeys(table, numRows / 10, "key000000100".length());
+
+    final ByteKeyRange middleRange = ByteKeyRange.of(startKey, endKey);
+    List<Row> middleRows = filterToRange(testRows, middleRange);
+    runReadTest(defaultRead.withTableId(table).withKeyRanges(StaticValueProvider.of(Collections.singletonList(middleRange))), middleRows);
+
+    assertThat(middleRows, allOf(hasSize(lessThan(numRows)), hasSize(greaterThan(0))));
   }
 
   /** Tests reading three key ranges with one read. */
@@ -614,6 +639,31 @@ public class BigtableIOTest {
         defaultRead.withTableId(table).withRowFilter(filter), Lists.newArrayList(filteredRows));
   }
 
+  /** Tests reading rows using a filter provided through ValueProvider. */
+  @Test
+  public void testReadingWithRuntimeParameterizedFilter() throws Exception {
+    final String table = "TEST-FILTER-TABLE";
+    final int numRows = 1001;
+    List<Row> testRows = makeTableData(table, numRows);
+    String regex = ".*17.*";
+    final KeyMatchesRegex keyPredicate = new KeyMatchesRegex(regex);
+    Iterable<Row> filteredRows =
+        testRows.stream()
+            .filter(
+                input -> {
+                  verifyNotNull(input, "input");
+                  return keyPredicate.apply(input.getKey());
+                })
+            .collect(Collectors.toList());
+
+    RowFilter filter =
+        RowFilter.newBuilder().setRowKeyRegexFilter(ByteString.copyFromUtf8(regex)).build();
+    service.setupSampleRowKeys(table, 5, 10L);
+
+    runReadTest(
+        defaultRead.withTableId(table).withRowFilter(StaticValueProvider.of(filter)),
+        Lists.newArrayList(filteredRows));
+  }
   /** Tests dynamic work rebalancing exhaustively. */
   @Test
   public void testReadingSplitAtFractionExhaustive() throws Exception {
@@ -626,10 +676,11 @@ public class BigtableIOTest {
 
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null,
-            Arrays.asList(service.getTableRange(table)),
-            null);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(
+                    StaticValueProvider.of(Collections.singletonList(service.getTableRange(table))))
+                .build(), null);
     assertSplitAtFractionExhaustive(source, null);
   }
 
@@ -645,10 +696,11 @@ public class BigtableIOTest {
 
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null,
-            Arrays.asList(service.getTableRange(table)),
-            null);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(
+                    StaticValueProvider.of(Collections.singletonList(service.getTableRange(table))))
+                .build(), null);
     // With 0 items read, all split requests will fail.
     assertSplitAtFractionFails(source, 0, 0.1, null /* options */);
     assertSplitAtFractionFails(source, 0, 1.0, null /* options */);
@@ -679,10 +731,8 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            Arrays.asList(ByteKeyRange.ALL_KEYS),
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),BigtableReadOptions.builder()
+            .setKeyRanges(ALL_KEY_RANGE).build(), null /*size*/);
     List<BigtableSource> splits =
         source.split(numRows * bytesPerRow / numSamples, null /* options */);
 
@@ -747,10 +797,9 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            keyRanges,
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider.of(keyRanges)).build(), null /*size*/);
 
     List<BigtableSource> splits = new ArrayList<>();
     for (ByteKeyRange range : keyRanges) {
@@ -798,10 +847,9 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            keyRanges,
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider.of(keyRanges)).build(), null /*size*/);
 
     List<BigtableSource> splits = new ArrayList<>();
     for (ByteKeyRange range : keyRanges) {
@@ -840,11 +888,8 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            Arrays.asList(ByteKeyRange.ALL_KEYS),
-            null /*size*/);
-
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder().setKeyRanges(ALL_KEY_RANGE).build(), null /*size*/);
     List<BigtableSource> splits = new ArrayList<>();
     List<ByteKeyRange> keyRanges =
         Arrays.asList(
@@ -913,16 +958,14 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            keyRanges,
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider.of(keyRanges)).build(), null /*size*/);
     BigtableSource referenceSource =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            ImmutableList.of(service.getTableRange(table)),
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider.of(Collections.singletonList(service.getTableRange(table)))).build(), null /*size*/);
     List<BigtableSource> splits = // 10,000
         source.split(numRows * bytesPerRow / numSamples, null /* options */);
 
@@ -947,10 +990,8 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            Arrays.asList(ByteKeyRange.ALL_KEYS),
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder().setKeyRanges(ALL_KEY_RANGE).build(), null /*size*/);
     List<BigtableSource> splits = source.split(numRows * bytesPerRow / numSplits, null);
 
     // Test num splits and split equality.
@@ -990,16 +1031,16 @@ public class BigtableIOTest {
     // Generate source and split it.
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            keyRanges,
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider.of(keyRanges)).build(),
             null /*size*/);
     BigtableSource referenceSource =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null /*filter*/,
-            ImmutableList.of(service.getTableRange(table)),
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setKeyRanges(StaticValueProvider
+                    .of(ImmutableList.of(service.getTableRange(table)))).build(), null /*size*/);
     List<BigtableSource> splits = source.split(numRows * bytesPerRow / numSplits, null);
 
     // Test num splits and split equality.
@@ -1025,10 +1066,10 @@ public class BigtableIOTest {
         RowFilter.newBuilder().setRowKeyRegexFilter(ByteString.copyFromUtf8(".*17.*")).build();
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            filter,
-            Arrays.asList(ByteKeyRange.ALL_KEYS),
-            null /*size*/);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder()
+                .setRowFilter(StaticValueProvider.of(filter))
+                .setKeyRanges(ALL_KEY_RANGE).build(), null /*size*/);
     List<BigtableSource> splits = source.split(numRows * bytesPerRow / numSplits, null);
 
     // Test num splits and split equality.
@@ -1058,7 +1099,7 @@ public class BigtableIOTest {
 
     assertThat(displayData, hasDisplayItem("rowFilter", rowFilter.toString()));
 
-    assertThat(displayData, hasDisplayItem("keyRange 0", keyRange.toString()));
+    assertThat(displayData, hasDisplayItem("keyRanges", "[ByteKeyRange{startKey=[], endKey=[abcd]}]"));
 
     // BigtableIO adds user-agent to options; assert only on key and not value.
     assertThat(displayData, hasDisplayItem("bigtableOptions"));
@@ -1100,7 +1141,7 @@ public class BigtableIOTest {
             .withBigtableOptions(BIGTABLE_OPTIONS)
             .withProjectId(options.getBigtableProject())
             .withInstanceId(options.getBigtableInstanceId())
-            .withTableId(options.getBigtableTableId());
+            .withTableId(options.getBigtableTableId())
     DisplayData displayData = DisplayData.from(read);
     assertThat(
         displayData,
@@ -1360,10 +1401,8 @@ public class BigtableIOTest {
 
     BigtableSource source =
         new BigtableSource(
-            config.withTableId(ValueProvider.StaticValueProvider.of(table)),
-            null,
-            Arrays.asList(ByteKeyRange.ALL_KEYS),
-            null);
+            config.withTableId(StaticValueProvider.of(table)),
+            BigtableReadOptions.builder().setKeyRanges(ALL_KEY_RANGE).build(), null);
 
     BoundedReader<Row> reader = source.createReader(TestPipeline.testingPipelineOptions());